+++ amd_openvx_extensions/amd_rpp/include/internal_rpp.h
@@ -43,6 +43,7 @@ THE SOFTWARE.
 #include<iostream>
 #include<algorithm>
 #include<functional>
+#include<map>
 
 using namespace std;
 
@@ -68,7 +69,22 @@ enum vxTensorLayout {
     VX_NHWC = 0,
     VX_NCHW = 1,
     VX_NFHWC = 2,
-    VX_NFCHW = 3
+    VX_NFCHW = 3,
+    VX_NHW = 4,     // Audio/2D layout
+    VX_NFT = 5,     // Frequency major, Used for Spectrogram/MelFilterBank
+    VX_NTF = 6      // Time major, Used for Spectrogram/MelFilterBank
+};
+
+const std::map<vxTensorLayout, RpptLayout> tensorLayoutMapping = {
+    {vxTensorLayout::VX_NHWC, RpptLayout::NHWC},
+    {vxTensorLayout::VX_NCHW, RpptLayout::NCHW},
+    {vxTensorLayout::VX_NFHWC, RpptLayout::NHWC},
+    {vxTensorLayout::VX_NFCHW, RpptLayout::NCHW},
+#if RPP_AUDIO
+    {vxTensorLayout::VX_NHW, RpptLayout::NHW},
+    {vxTensorLayout::VX_NFT, RpptLayout::NFT},
+    {vxTensorLayout::VX_NTF, RpptLayout::NTF}
+#endif
 };
 
 //! Brief The utility functions
@@ -76,6 +92,8 @@ vx_node createNode(vx_graph graph, vx_enum kernelEnum, vx_reference params[], vx
 vx_status createRPPHandle(vx_node node, vxRppHandle ** pHandle, Rpp32u batchSize, Rpp32u deviceType);
 vx_status releaseRPPHandle(vx_node node, vxRppHandle * handle, Rpp32u deviceType);
 void fillDescriptionPtrfromDims(RpptDescPtr &descPtr, vxTensorLayout layout, size_t *tensorDims);
+void fillGenericDescriptionPtrfromDims(RpptGenericDescPtr &genericDescPtr, vxTensorLayout layout, size_t *maxTensorDims);
+void fillAudioDescriptionPtrFromDims(RpptDescPtr &descPtr, size_t *maxTensorDims, vxTensorLayout layout = vxTensorLayout::VX_NHW);
 RpptDataType getRpptDataType(vx_enum dataType);
 
 class Kernellist